Refactor home screen associated files
