+++ src/components/main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep3Screen.jsx
@@ -13,11 +13,11 @@ import {
 } from "react-native";
 import { Button, Snackbar } from "react-native-paper";
 import Animated, { BounceIn, FadeIn, FadeOut } from "react-native-reanimated";
-import AlgoquantApiContext from "../../../constants/ApiContext";
-import { THEME } from "../../../constants/Theme";
-import { snackbarCleanUp } from "../../../helpers/snackbarCleanup";
-import SnackbarContent from "../../reusable_components/SnackbarContent";
-import CreateInvestorStockSearch from "../../single_use_components/CreateInvestorStockSearch";
+import AlgoquantApiContext from "../../../../../constants/ApiContext";
+import { THEME } from "../../../../../constants/Theme";
+import { snackbarCleanUp } from "../../../../../helpers/snackbarCleanup";
+import SnackbarContent from "../../../../reusable_components/SnackbarContent";
+import CreateInvestorStockSearch from "../../../../single_use_components/CreateInvestorStockSearch";
 
 export default function CreateInvestorAlgorithmicStep3Screen(props) {
   const { investorObject } = props.route.params;
